chore(ektorp): update used ektorp to 1.4.4
