+++ NOTICE-dependencies
@@ -53,7 +53,7 @@ log4j-1.2.17-sources.jar
 lucene-core-2.3.1-sources.jar
 mail-1.4-sources.jar
 opencsv-2.3-sources.jar
-org.ektorp-1.4.2-sources.jar
+org.ektorp-1.4.4-sources.jar
 oro-2.0.8-sources.jar
 poi-3.9-sources.jar
 poi-ooxml-3.9-sources.jar
